+++ spec/grid/base/grid.spec.ts
@@ -602,7 +602,21 @@ describe('Grid base module', () => {
         beforeAll((done: Function) => {
             gridObj = createGrid(
                 {
-                    dataSource: filterData,
+                    dataSource: [{
+                        OrderID: 10248, CustomerID: 'VINET', EmployeeID: 5, OrderDate: new Date("07 12 1996 02:00:23"),
+                        ShipName: 'Vins et alcools Chevalier', ShipCity: 'Reims', ShipAddress: '59 rue de l Abbaye',
+                        ShipRegion: 'CJ', ShipPostalCode: '51100', ShipCountry: 'France', Freight: 32.38, Verified: !0
+                    },
+                    {
+                        OrderID: 10249, CustomerID: 'TOMSP', EmployeeID: 6, OrderDate: new Date("07 12 1996 00:03:23"),
+                        ShipName: 'Toms Spezialitäten', ShipCity: 'Münster', ShipAddress: 'Luisenstr. 48',
+                        ShipRegion: 'CJ', ShipPostalCode: '44087', ShipCountry: 'Germany', Freight: 11.61, Verified: !1
+                    },
+                    {
+                        OrderID: 10250, CustomerID: 'HANAR', EmployeeID: 4, OrderDate: new Date("07 12 1996 00:00:23"),
+                        ShipName: 'Hanari Carnes', ShipCity: 'Rio de Janeiro', ShipAddress: 'Rua do Paço, 67',
+                        ShipRegion: 'RJ', ShipPostalCode: '05454-876', ShipCountry: 'Brazil', Freight: 65.83, Verified: !0
+                    }],
                     allowFiltering: true,
                     allowPaging: true,
                     filterSettings: { type: 'Menu' },
@@ -618,13 +632,13 @@ describe('Grid base module', () => {
                 done);
         });
         it('update particular cell', () => {
-            gridObj.setCellValue(10248, 'CustomerID', 'new value');
-            let selRow: any = gridObj.contentModule.getRows()[0];
+            gridObj.setCellValue(10249, 'CustomerID', 'new value');
+            let selRow: any = gridObj.contentModule.getRows()[1];
             expect((<any>selRow).data.CustomerID).toEqual('new value');
         });
         it('update particular row', () => {
-            gridObj.setRowData(10249, { OrderID: 1249, CustomerID: 'new value', CustomerName: 'accc' });
-            let selRow: any = gridObj.contentModule.getRows()[1];
+            gridObj.setRowData(10250, { OrderID: 1249, CustomerID: 'new value', CustomerName: 'accc' });
+            let selRow: any = gridObj.contentModule.getRows()[2];
             expect((<any>selRow).data.CustomerID).toEqual('new value');
         });
         afterAll(() => {